bot/requirements.txt
# =============================================================================
# Bot Discord Minecraft - Dépendances Python
# =============================================================================

# -----------------------------------------------------------------------------
# Discord Bot Framework
# -----------------------------------------------------------------------------
discord.py>=2.3.2
discord-py-interactions>=5.12.0

# -----------------------------------------------------------------------------
# Async & Concurrency
# -----------------------------------------------------------------------------
asyncio>=3.4.3
aiohttp>=3.9.1
aiofiles>=23.2.1
aiocache>=0.12.2

# -----------------------------------------------------------------------------
# Database - PostgreSQL
# -----------------------------------------------------------------------------
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.23
alembic>=1.13.1
greenlet>=3.0.2

# -----------------------------------------------------------------------------
# Database - Redis
# -----------------------------------------------------------------------------
redis>=5.0.1
aioredis>=2.0.1

# -----------------------------------------------------------------------------
# API & Web
# -----------------------------------------------------------------------------
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
httpx>=0.26.0
websockets>=12.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# -----------------------------------------------------------------------------
# Docker Integration
# -----------------------------------------------------------------------------
docker>=7.0.0
aiodocker>=0.21.0

# -----------------------------------------------------------------------------
# Minecraft Protocol & RCON
# -----------------------------------------------------------------------------
mcrcon>=0.7.0
mcstatus>=11.1.0
nbt>=1.5.1

# -----------------------------------------------------------------------------
# Configuration & Environment
# -----------------------------------------------------------------------------
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tomli>=2.0.1

# -----------------------------------------------------------------------------
# Logging & Monitoring
# -----------------------------------------------------------------------------
loguru>=0.7.2
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
sentry-sdk>=1.38.0

# -----------------------------------------------------------------------------
# Security
# -----------------------------------------------------------------------------
cryptography>=41.0.7
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
itsdangerous>=2.1.2

# -----------------------------------------------------------------------------
# Task Scheduling
# -----------------------------------------------------------------------------
apscheduler>=3.10.4
celery>=5.3.4

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
orjson>=3.9.10
msgpack>=1.0.7
numpy>=1.26.2
pandas>=2.1.4

# -----------------------------------------------------------------------------
# CLI & Utilities
# -----------------------------------------------------------------------------
click>=8.1.7
rich>=13.7.0
typer>=0.9.0
tabulate>=0.9.0

# -----------------------------------------------------------------------------
# File & Image Processing
# -----------------------------------------------------------------------------
pillow>=10.1.0
aiopath>=0.7.7

# -----------------------------------------------------------------------------
# Date & Time
# -----------------------------------------------------------------------------
python-dateutil>=2.8.2
arrow>=1.3.0
pytz>=2023.3.post1

# -----------------------------------------------------------------------------
# Validation & Parsing
# -----------------------------------------------------------------------------
email-validator>=2.1.0.post1
validators>=0.22.0
parse>=1.20.0

# -----------------------------------------------------------------------------
# HTTP Client Features
# -----------------------------------------------------------------------------
requests>=2.31.0
urllib3>=2.1.0

# -----------------------------------------------------------------------------
# Template Engine
# -----------------------------------------------------------------------------
jinja2>=3.1.2

# -----------------------------------------------------------------------------
# Backup & Compression
# -----------------------------------------------------------------------------
py7zr>=0.20.8
rarfile>=4.1

# -----------------------------------------------------------------------------
# System Monitoring
# -----------------------------------------------------------------------------
psutil>=5.9.7
